Cleanup, rename, simplify usb query
