+++ firmware/src/core_m4/CMakeLists.txt
@@ -95,6 +95,7 @@ add_executable(
   ${FWDIR}/src/usb/device_midi/usbd_midi.c
   ${FWDIR}/src/usb/device_midi/usbd_midi_desc.c
   ${FWDIR}/src/usb/usb_midi_device.cc
+  ${FWDIR}/src/usb/usbh_midi_jacks.cc
 
   ${USBDEV_LIBDIR}/Core/Src/usbd_core.c
   ${USBDEV_LIBDIR}/Core/Src/usbd_ctlreq.c
