+++ firmware/src/usb/usbh_midi_jacks.cc
@@ -174,3 +174,29 @@ void scan_all_descriptors(USBH_HandleTypeDef const *phost) {
 		}
 	}
 }
+
+void count_midi_jacks(USBH_HandleTypeDef const *phost, uint8_t *num_in_jacks, uint8_t *num_out_jacks) {
+	uint8_t in_jacks = 0;
+	uint8_t out_jacks = 0;
+
+	unsigned desc_total_length = phost->device.CfgDesc.wTotalLength;
+	auto *pdesc = reinterpret_cast<USBH_DescHeader_t const *>(phost->device.CfgDesc_Raw);
+	uint16_t ptr = USB_LEN_CFG_DESC;
+
+	while (ptr < desc_total_length) {
+		pdesc = USBH_GetNextDesc((uint8_t *)(void *)pdesc, &ptr);
+
+		if (pdesc->bDescriptorType == 0x24) { // Class-specific MIDIStreaming descriptor
+			uint8_t subtype = ((uint8_t const *)pdesc)[2];
+			if (subtype == 0x02) // MIDI_IN_JACK
+				in_jacks++;
+			else if (subtype == 0x03) // MIDI_OUT_JACK
+				out_jacks++;
+		}
+	}
+
+	if (num_in_jacks)
+		*num_in_jacks = in_jacks;
+	if (num_out_jacks)
+		*num_out_jacks = out_jacks;
+}
